Fixed a bunch of stuff related to appending, added more toString defs, fixed array cast message, better working on custom runtime test

+++ arsd-webassembly/core/internal/utf.d
@@ -21,7 +21,7 @@
 
 module core.internal.utf;
 
-extern (C) void onUnicodeError( string msg, size_t idx, string file = __FILE__, size_t line = __LINE__ ) @trusted
+extern (C) void onUnicodeError( string msg, size_t idx, string file = __FILE__, uint line = __LINE__ ) @trusted
 {
     _d_assert_msg("onUnicodeError: "~msg, file, line);
 }
